kernel: Add new platform EN75xx and device SmartFiber XP8421-B
EcoNet EN75xx is a big endian MIPS platform used in XPON (fiber),
DSL, and SIM (3g/4g) applications. Complete GPL vender SDKs exist
for this platform, but are based on Linux 2.6.

The SmartFiber XP8421-B is a fiber modem which is available for
$20 online and has 512MB of memory, 256MB of SPI NAND flash and
2 USB 2.0 ports in addition to ethernet, wifi and XPON.

Many of the drivers for this SoC bear little resemblence to
anything currently supported in OpenWRT or Linux upstream, so
there is no realistic timeline for reaching a feature complete port.
The aim of this patch is to provide the foundation (INTC, timer,
SPI, NAND, image format) to be able to flash and boot to a shell.

I'm not getting paid for this, I'm just a hobbyist, and this is
my first contribution. So any advice and guidance to help get this
landed will be most appreciated.

The XP8421-B, and apparently many other devices of this platform,
use a dual-image layout. I have chosen to reuse this to support
dual-boot between OpenWRT and the factory firmware. Certain design
decisions were made with the goal of not overwriting data that is
used by the factory OS.

Flashing instructions (from bootloader):
1. Build and then locate the squashfs-tclinux.trx image file
2. Get the length of that file in hex:
	`printf '%X\n' "$(stat -c%s the-file-squashfs-tclinux.trx)"`
3. Connect to device with xmodem capability, e.g.
	`picocom --send-cmd lsx -vv -b 115200 /dev/ttyUSB0`
4. Switch device on and press a key within 3 seconds
5. Enter bootloader username and password: telecomadmin nE7jA%5m
6. Type: `xmdm 80020000 <file length hex>`
7. Quickly start xmodem and send the file, in picocom that is
	`ctrl+a ctrl+s <paste-the-file-name> enter`
    If the transfer fails to start, wait 30 seconds to a minute for
	the bootloader prompt to return and then try the command again.
8. Once the transfer has completed successfully, type the following
    `flash 80000 80020000 <file length hex>`
9. Type `go` or simply restart the device to boot into OpenWRT

Signed-off-by: Caleb James DeLisle <cjd@cjdns.fr>